requirements.txt
lxml>=4.2.5
pycryptodome>=3.7.2
requests>=2.21.0